refactor(audit): split shared audit helpers by domain
